requirements.txt
asyncio
python-dotenv
telethon
wand
cryptg
